add separate default components for badge and title
