+++ src/components/content-types/TopicList.tsx
@@ -4,7 +4,6 @@ import { ContentProps, EditableContentProps } from "../Content"
 import { useDocument, useHandle, useRepo } from "automerge-repo-react-hooks"
 import "./TextContent.css"
 import { DocCollection, DocHandle, DocumentId } from "automerge-repo"
-import ContentList, { ContentListInList } from "./ContentList"
 import { BoardDoc, BoardDocCard } from "./board"
 import { parseDocumentLink, PushpinUrl } from "../pushpin-code/Url"
 import { TextDoc } from "./TextContent"
