+++ src/components/content-types/defaults/DefaultInTitle.tsx
@@ -1,0 +1,46 @@
+import React from "react"
+import * as ContentTypes from "../../pushpin-code/ContentTypes"
+import { ContentProps } from "../../Content"
+import TitleWithSubtitle from "../../ui/TitleWithSubtitle"
+import { useDocument } from "automerge-repo-react-hooks"
+import { useLastSeenHeads } from "../../pushpin-code/Changes"
+import { createDocumentLink } from "../../pushpin-code/Url"
+import { readWithSchema } from "../../../lenses"
+import { HasTitle } from "../../../lenses/HasTitle"
+
+interface Doc {
+  title?: string
+}
+
+export default function DefaultInTitle(props: ContentProps) {
+  const { url, documentId } = props
+  const [rawDoc] = useDocument<Doc>(documentId)
+  const { type } = props
+  const lastSeenHeads = useLastSeenHeads(createDocumentLink(type, documentId))
+
+  if (!rawDoc) {
+    return null
+  }
+
+  const doc = readWithSchema({
+    doc: rawDoc,
+    type,
+    lastSeenHeads,
+    schema: "HasTitle",
+  }) as HasTitle
+
+  return (
+    <TitleWithSubtitle
+      title={doc.title}
+      titleEditorField={doc.titleEditorField ?? ""}
+      subtitle={doc.subtitle}
+      documentId={documentId}
+      editable={doc.titleEditorField !== null}
+    />
+  )
+}
+
+ContentTypes.registerDefault({
+  component: DefaultInTitle,
+  context: "title",
+})
